requirements.txt
# Requirements needed for the react-lms app

django-webpack-loader==0.5.0
